+++ docs/CHANGELOG.rst
@@ -6,6 +6,7 @@
     * Fixed #442 no special characters allowed for `license_key`, `license_name` and `license_expression`
     * Fixed #446 Better error handling
     * Add support to collect redistributable sources #22
+    * Handle trailing spaces in field names during `transform` #456
 
 2020-08-11
     Release 5.0.0